resource ref children (#1798)
* resource ref children

* working resource refs

* big cleanup

* even more cleanup and reorg

* fix a typo

* update docs, rename ResourceRefs → ResourceRef

* provide a default empty enum

* add dynamic enum docs

* fix value, remove value on hide

* bump image version

* fix mock

* Update docs/extensibility/form-widgets.md
