+++ core-ui/src/components/Extensibility/ExtensibilityCreate.js
@@ -3,18 +3,19 @@ import { createStore } from '@ui-schema/ui-schema';
 import { createOrderedMap } from '@ui-schema/ui-schema/Utils/createMap';
 import Immutable from 'immutable';
 import pluralize from 'pluralize';
+import { useTranslation } from 'react-i18next';
 
 import { ResourceForm } from 'shared/ResourceForm';
 import { useMicrofrontendContext } from 'shared/contexts/MicrofrontendContext';
-import { createTemplate } from './helpers';
-
-import { ResourceSchema } from './ResourceSchema';
 import { useNotification } from 'shared/contexts/NotificationContext';
-import { useTranslation } from 'react-i18next';
 import { Spinner } from 'shared/components/Spinner/Spinner';
 import { useGetSchema } from 'hooks/useGetSchema';
 import { prettifyKind } from 'shared/utils/helpers';
-import { prepareSchemaRules } from './SchemaRulesInjector';
+
+import { ResourceSchema } from './ResourceSchema';
+import { createTemplate } from './helpers';
+import { VarStoreContextProvider } from './contexts/VarStore';
+import { prepareSchemaRules } from './helpers/prepareSchemaRules';
 
 export function ExtensibilityCreate({
   formElementRef,
@@ -26,7 +27,6 @@ export function ExtensibilityCreate({
   toggleFormFn,
   resourceName,
 }) {
-  const [varStore, setVarStore] = useState({});
   const { namespaceId: namespace } = useMicrofrontendContext();
   const notification = useNotification();
   const { t } = useTranslation();
@@ -96,44 +96,42 @@ export function ExtensibilityCreate({
   if (loading) return <Spinner />;
 
   return (
-    <ResourceForm
-      pluralKind={resourceType}
-      singularName={pluralize(resourceName || prettifyKind(resource.kind), 1)}
-      resource={resource}
-      setResource={updateResource}
-      formElementRef={formElementRef}
-      createUrl={resourceUrl}
-      setCustomValid={setCustomValid}
-      onlyYaml={!schema}
-      initialResource={initialResource}
-      afterCreatedFn={afterCreatedFn}
-    >
-      <ResourceSchema
-        simple
-        key={api.version}
-        schema={errorOpenApi ? {} : schema}
-        schemaRules={simpleRules}
+    <VarStoreContextProvider>
+      <ResourceForm
+        pluralKind={resourceType}
+        singularName={pluralize(resourceName || prettifyKind(resource.kind), 1)}
         resource={resource}
-        store={store}
-        setStore={setStore}
-        onSubmit={() => {}}
-        path={general?.urlPath || ''}
-        varStore={varStore}
-        setVarStore={setVarStore}
-      />
-      <ResourceSchema
-        advanced
-        key={api.version}
-        schema={errorOpenApi ? {} : schema}
-        schemaRules={advancedRules}
-        resource={resource}
-        store={store}
-        setStore={setStore}
-        path={general?.urlPath || ''}
-        varStore={varStore}
-        setVarStore={setVarStore}
-      />
-    </ResourceForm>
+        setResource={updateResource}
+        formElementRef={formElementRef}
+        createUrl={resourceUrl}
+        setCustomValid={setCustomValid}
+        onlyYaml={!schema}
+        initialResource={initialResource}
+        afterCreatedFn={afterCreatedFn}
+      >
+        <ResourceSchema
+          simple
+          key={api.version}
+          schema={errorOpenApi ? {} : schema}
+          schemaRules={simpleRules}
+          resource={resource}
+          store={store}
+          setStore={setStore}
+          onSubmit={() => {}}
+          path={general?.urlPath || ''}
+        />
+        <ResourceSchema
+          advanced
+          key={api.version}
+          schema={errorOpenApi ? {} : schema}
+          schemaRules={advancedRules}
+          resource={resource}
+          store={store}
+          setStore={setStore}
+          path={general?.urlPath || ''}
+        />
+      </ResourceForm>
+    </VarStoreContextProvider>
   );
 }
