+++ core-ui/src/components/Extensibility/components-form/index.js
@@ -12,7 +12,9 @@ import { PluginSimpleStack } from '@ui-schema/ui-schema/PluginSimpleStack';
 import { ValidityReporter } from '@ui-schema/ui-schema/ValidityReporter';
 import { validators } from '@ui-schema/ui-schema/Validators/validators';
 
-import { SchemaRulesInjector } from '../SchemaRulesInjector';
+import { SchemaRulesInjector } from '../plugins/SchemaRulesInjector';
+import { EnumHandler } from '../plugins/EnumHandler';
+import { VisibilityHandler } from '../plugins/VisibilityHandler';
 
 import { StringRenderer } from './StringRenderer';
 import { NumberRenderer } from './NumberRenderer';
@@ -34,6 +36,8 @@ const pluginStack = [
   DependentHandler,
   ConditionalHandler,
   SchemaRulesInjector,
+  EnumHandler,
+  VisibilityHandler,
   PluginSimpleStack,
   ValidityReporter,
 ];
@@ -77,7 +81,7 @@ export const widgets = {
     KeyValuePair: KeyValuePairRenderer,
     CodeEditor: MonacoRenderer,
     FormGroup: CollapsibleRenderer,
-    ResourceRefs: ResourceRefRender,
+    ResourceRef: ResourceRefRender,
     Resource: ResourceRenderer,
   },
 };
