+++ core-ui/src/components/Extensibility/helpers/jsonataWrapper.js
