+++ docs/extensibility/assets/form-widgets/ResourceRef.png
